requirements.txt
pycryptodome>=3.19,<4
zstandard>=0.22,<1
mcp>=1.0,<2
pilk>=0.2
pyinstaller>=6.0
